Remove stopping condition for C(SProd) and C(Exp) (#1247)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Follow-up to https://github.com/PennyLaneAI/pennylane/pull/8133, where
PennyLane can now decompose SProd and Exp, and it is no longer necessary
for Lightning to convert C(SProd) and C(Exp) to qubit unitary.

**Description of the Change:**
Remove stopping condition for C(SProd) and C(Exp), which were previously
added in https://github.com/PennyLaneAI/pennylane-lightning/pull/1120 to
solve https://github.com/PennyLaneAI/pennylane-lightning/issues/900

**Benefits:**
No need to convert those gates to matrix, and let PennyLane decompose as
it would for any plugin devices.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-98636]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>