Upgrade CI to Python 3.12, CIBuildWheel 3.1.4 + Fix docker build (#1248)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently most of our CIs use Python 3.11 - here we upgrade to 3.12.
Also CIBuildWheel is currently pinned to v2.21.0.
Also our docker builds have been
[failing](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17422611595)
for some time since it uses Python 3.10 (default from ubuntu 22), which
is no longer supported by PL.

**Description of the Change:**
Upgrade CIs to use Python3.12 wherever found.
CIBuildWheel now uses latest 3.1.4

For the dockerfile, we use deadsnake to get Python3.12 for ubuntu22.

**Benefits:**
Docker pass again:
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17477483960

**Possible Drawbacks:**
The minimum supported version currently is 3.11, so there is the
possibility that some bugs for 3.11 might be missed - we do test some
3.11 wheels though.

**Related GitHub Issues:**

[sc-76184]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>