Provide an (updated) qubit manager for lightning
The existing setup pulled in the qubit manager from the catalyst runtime
repo, but that one was designed for the StateVectorLQubitDynamic class,
which does not match current lightning device behaviour. For instance,
upon freeing a qubit the old qubit manager would decrement all device
IDs in the map in order to preserve contiguity of device IDs, but this
assumes the device is actually shrinking the statevector upon dealloc,
which lightning devices currently do not do (= bug).

Instead, the code is adapted to be more suitable for the current device
behaviour, and maintains a set of freed qubits that can be reinitialized
for new allocations.