QuantumDevice interface updates to lightning for dynamic allocation (#1254)
The current implementation of Catalyst's `QuantumDevice` interface in
lightning has some issues, in that it doesn't follow the required
semantics. Specifically, around partial qubit deallocation and their
effect on other interface functions, like `GetNumQubits` and MP
functions.

In preparation for the upcoming dynamic allocation feature in PennyLane,
these inconsistencies need to be fixed. We are also replacing the
`ReleaseAllQubits` function with a `ReleaseQubits([ids...])`, which was
a longstanding inconsistency between the Runtime C-API and the
QuantumDevice interface.

Note: This PR will be merged, then the corresponding Catalyst PR will be
merged. We will then have a [new separate
PR](https://github.com/PennyLaneAI/pennylane-lightning/pull/1258) to
undo the changes to our CI to track their main branch again.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Joseph Lee <joseph.lee@xanadu.ai>
Co-authored-by: Joseph Lee <40768758+josephleekl@users.noreply.github.com>
Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>