Update Kokkos installation instruction (#1256)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
When installing Kokkos, we require `Kokkos_ENABLE_COMPLEX_ALIGN=OFF`,
otherwise std::complex cannot be cast to kokkos::complex. This setting
is set when we let lightning handle compiling Kokkos, but in the
installation guide for using an external Kokkos, this is not included.

**Description of the Change:**
Add setting in installation guide.

**Benefits:**
Successful build.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-99167]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Jake Zaia <23638795+jzaia18@users.noreply.github.com>
Co-authored-by: ANT0N <39093564+AntonNI8@users.noreply.github.com>