[Catalyst] Properly reset released qubits by measuring (#1260)
**Context:**
In #1254 , the Catalyst interfaces for lighting.qubit, kokkos and gpu
were updated to ready for the new dynamic allocation feature.
One specific part was the reuse. Namely, previously released qubits are
free to be reused in future allocations.

The `ReleaseQubit()` method essentially just records which qubits are in
the recycle pool, and future calls to `AllocateQubit()` method should
get a qubit from the pool and reset it to |0>.

However, the reset was just calling `collapse(0)`, which is usually
fine, but then the qubit in the recycle pool was in |1> state, i.e. had
no overlap with |0> at all, then the collapse would fail, because
essentially the state vector is now having zero amplitude on both |0>
and |1>, and normalization returns `nan`.

**Description of the Change:**
We properly reset by performing a measurement, and inserting an X gate
if |1> was measured.

**Benefits:**
Correct logic in qubit release and reuse for catalyst's dynamic
allocation feature.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Joseph Lee <joseph.lee@xanadu.ai>