Remove unnecessary std::move that will cause error in g++13 (#1263)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
In preparation for upgrading to GCC-13, there is a line that will cause
an error due to `-Wpessimizing-move`

```
/opt/actions-runner/_work/pennylane-lightning/pennylane-lightning/main/pennylane_lightning/core/utils/cuda_utils/cuGates_host.hpp:604:16: error: moving a temporary object prevents copy elision [-Werror=pessimizing-move]
604 | const auto rot{std::move(getRot<CFP_t>(phi, theta, omega))};
| ^~~
/opt/actions-runner/_work/pennylane-lightning/pennylane-lightning/main/pennylane_lightning/core/utils/cuda_utils/cuGates_host.hpp:604:16: note: remove ‘std::move’ call
```

The reason is that this `std::move` prevents Return Value Optimization.

**Description of the Change:**
Remove this std::move.

**Benefits:**
So that we can upgrade to GCC-13 smoothly
([PR](https://github.com/PennyLaneAI/pennylane-lightning/pull/1249/files#diff-e41e7d5f076e7742b8de760853126cfcfefad666132dc1c989eeaa4ce25f10e9))
- after release

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-100159]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>