Raise error for `normalize` when norm close to zero for LQ/`collapse` for LGPU (#1257)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
We have previously commented out the code to raise an error for LQ when
normalizing a SV with norm close to 0. This was done to have the same
behaviour as PennyLane for edge cases in MCM. However this scenario
should actually raise an error (and behaviour in PennyLane is also not
which is
[undecided](https://github.com/PennyLaneAI/pennylane/issues/6504) ), and
this is also required for updating the method to [dynamically allocate
qubits for
Catalyst](https://github.com/PennyLaneAI/pennylane-lightning/pull/1254)

**Description of the Change:**
Abort and raise error when normalizing with norm close to 0.

**Benefits:**
Unblocks dynamic re-allocation of qubits for Catalyst.

**Possible Drawbacks:**
Some code which can run in the past with close to zero norm (which
should have undefined result) e.g. with MCM now will be aborted.

**Related GitHub Issues:**

Previous discussion:
https://xanaduhq.slack.com/archives/CA89H1BAN/p1748366977551839

[latest/latest
](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17740522340)

[sc-99285]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>