support compiling Lightning GPU and Tensor from source with CUDA 13 (#1267)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently Lightning GPU and Tensor do not support building with CUDA 13.

**Description of the Change:**
Allow users to choose building Lightning GPU and Tensor with CUDA 13.
This can be enabled by:
- adding `PL_CUDA_VERSION` environment variable (or to `CMAKE_ARGS`), or
- auto detect CUDA version from NVCC, or
- default uses CUDA 12

Note: We are not pre-building wheels or performing CI tests with CUDA 13
yet - just enabling the possibility of building from source with CUDA
13.

Note: This has been tested with GCC-14.3.0

Note: The current version of Kokkos used (4.5.0) does not support
compiling with CUDA 13. This support is only added from
[v4.7.1](https://github.com/kokkos/kokkos/pull/8328). Therefore we are
not supporting building Lightning Kokkos with CUDA 13 yet.


**Benefits:**
Enables Lightning GPU and Tensor devices for systems with CUDA 13.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-100445]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>