PR to RC branch do not upload to TestPyPI (#1275)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Previously when we have a PR basing on the rc branch, when we use
`ci:build_wheel`, it automatically uploads to TestPyPI. We do not want
this behaviour, and only the PR of the rc branch itself to upload to
PyPI.

**Description of the Change:**
Fixed the above behaviour. 

Also revert change of updating test seed, which should be done in the
master branch.

Both of these changes are reverting/updating previous commits.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**