LQ CPU Python CI with release use tests from rc branch (#1280)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently when [testing the release
candidates](https://github.com/PennyLaneAI/pennylane-lightning/actions/workflows/compat-check-release-release.yml),
lqcpu python use lightning tests from the master branch, instead of the
rc branch. This was because the action is split into a build phase and a
test phase, and only during the build phase we checkout the rc branch,
that means during the test phase, it stlil pulls the master branch. That
means when tests are updated in the rc branch, they will not be
reflected during rc testing, and can cause
[failure](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18366101157).

**Description of the Change:**
Checkout rc branch for lqcpu testing for rc.

**Benefits:**

[updated test now
pass](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18379012402)
**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>