Fix new MCM test seed and skip for LT (#1281)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
A recently added tests (`reuse_with_mcm`) uses shots, and did not fix
the seed. Also this test should not be compatible with LT.

**Description of the Change:**
Skip MCM for LT, and fix seed for deterministic test result.

**Benefits:**
Remove flaky test.

**Possible Drawbacks:**

**Related GitHub Issues:**