Remove catalyst pin for stable test (#1284)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
We previously added a catalyst version pin for stable tests which uses
the older version of CAPI. This is no longer necessary with the new
releases.

**Description of the Change:**
Remove catalyst pin for stable tests. Without this removal, stable tests
will
[fail](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18539279261).

**Benefits:**
stable tests will pass again


[stable/stable](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18540216053)

[stable/latest](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18540230497)

[latest/stable](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18540234521)

[latest/latest](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18540235673)

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-101543]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>