Use GCC 13 for CI (#1249)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------


**Context:**
Note: we will merge this after 0.43 release

We want to upgrade our CIs from current GCC 11 to a more modern one (13)
which supports more C++-20 features.

**Description of the Change:**
Upgrade CIs to GCC 13.

Note: For CUDA wheel builds (x86 and aarch64), we can't upgrade since we
use CUDA toolkit 12.0, and the maximum supported GCC is 12.

**Benefits:**
Support more C++-20 features.

**Possible Drawbacks:**

**Related GitHub Issues:**



[latest/latest](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18692539901)

[stable/stable](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18692537920)

[docker](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/18687414873/job/53283733452)
[sc-92369]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Jake Zaia <23638795+jzaia18@users.noreply.github.com>