Remove wheel building for x86 macOS (#1287)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
[The PennyLane ecosystem will no longer support Intel MacOS platforms
for
v0.44](https://pennylane.ai/blog/2025/10/pennylane-release-0.43/#deprecations-and-breaking-changes-).
This PR prepares PennyLane-Lightning to stop producing Python wheels for
MacOS x86. If needed, MacOS x86 wheels can be built manually.

This change is needed to account for MacOS officially deprecating
support for Intel CPUs (see their [blog
post](https://github.blog/changelog/2025-09-19-github-actions-macos-13-runner-image-is-closing-down/#notice-of-macos-x86_64-intel-architecture-deprecation)
for more details).

Note: MacOS ARM wheels will require a minimum OS version of 14.0 for
continued use with v0.44 and newer.



**Description of the Change:**
Stop producing MacOS x86 wheels on CIs.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-101581]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>