Remove operator instantiation from test to obtain matrix (#1292)
**Context:**
A test uses an operator instantiation within a QNode to compute a sparse
matrix. This should not be done within a QNode
and will surface as a problem once
https://github.com/PennyLaneAI/pennylane/pull/8131 is merged.

**Description of the Change:**
Compute the matrix outside of the QNode instead.

**Benefits:**
Code quality, and compliance with new queuing behaviour that will be
created by the mentioned PR.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
N/A

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Jake Zaia <23638795+jzaia18@users.noreply.github.com>