Update MacOS runner to 14 (#1304)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
MacOS 13 runners are deprecated and have to be updated to 14.

With this update, we observe the same Py3.14 networkx error as aarch64
wheels:

https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/20037251654/job/57461938825?pr=1304
We implement the same patch here.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-105494]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>