Cache intermediate states in adjoint/VJP to eliminate redundant forward pass in backward (#1312)
**Context:**
Cache intermediate states in adjoint/VJP to eliminate redundant forward
pass in backward.

Currently, in the backward pass (adjoint-jacobian, vjp), there are
redundant calls to the forward pass to get the final state. This new
cache approach will help to avoid redundant computations by storing the
final state after the forward pass and reusing it in the backward pass.
This works with single or multiple backward calls after a single forward
pass. The current cache is per-device, so if multiple devices are used,
each device will have its own cache and will be freed when the device is
deallocated. This boosts the performance of backward execution on CPUs
significantly by avoiding extra forward executions of the same circuit.
For GPU devices, the memory walls and high compute demands make thie
caching ineffective at scale, so it is disabled by default for GPU
devices.


**Benefits:**
- Checked the backward/fordward ratio of referenced example (PR 1311)
for LQ (24, 4, 2), this PR reduces the wall-time ratio from `8.1` ->
`5.6`.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-106746]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: tomlqc <thomas.germain@xanadu.ai>
Co-authored-by: Jake Zaia <23638795+jzaia18@users.noreply.github.com>